requirements.txt
Flask
qrcode[pil]
werkzeug
gunicorn
python-telegram-bot==13.15
requests